🚨(crawlers): More tests
More tests for parsing methods, used by crawlers.
